Solutions for 01, 02, and 03 of flex exercises
Exercise 3 was a major doozy for me. Firstly, I struggle with fixing
existing code. I feel like if I built the stylesheet and the HTML from
the ground up, it would go much easier for me. Secondly, I struggled
a lot with getting space between the items inside the left and right
box. I understood that I needed to place the logo and list in their own
box and the same for the remaining elements so that they could be
divided into left and right sides using justify-content: space-between
in the header div. But I didn't understand why I couldn't use
justify-content again in the the child divs. It seems that
justify-content in the parent div uses all of the space initially,
"compressing" the items in the child divs, leaving no space for them
to be separated using justify-contents again. This is where gap comes
in, where you can force space between the elements, ignoring the parent
div's justify-content.