Solution for 04 of flex exercises
This one was really easy for me. I did not struggle at all or need to
look at any resources for help, from the past lessons or from Stack
Overflow. It was like I was in the zone for the first time and the code
just flowed out of my fingers. It reassured me that I am actually
learning and gave me confidence in my developing skills.